Upgraded to Eventuate Local 0.2.0.RELEASE

+++ _build-and-test-all.sh
@@ -19,6 +19,12 @@ else
   ${DOCKER_COMPOSE?} rm -v --force
 fi
 
+NO_RM=false
+
+if [ "$1" = "--no-rm" ] ; then
+  NO_RM=true
+  shift
+fi
 
 ${DOCKER_COMPOSE?} up -d mongodb
 
@@ -54,5 +60,7 @@ set -e
 
 ./gradlew -a $* :e2e-test:cleanTest :e2e-test:test -P ignoreE2EFailures=false
 
-${DOCKER_COMPOSE?} stop
-${DOCKER_COMPOSE?} rm -v --force
+if [ $NO_RM = false ] ; then
+  ${DOCKER_COMPOSE?} stop
+  ${DOCKER_COMPOSE?} rm -v --force
+fi
